backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-socketio[asyncio]==5.10.0
aiofiles==23.2.1